refactor(ui): extract history view into shared prop-driven component
Move the promotion History view out of the dashboard's HistoryPage into a
reusable HistoryView component in components-lib, so the ArgoCD app-view
extension can render it without depending on a router.

- HistoryView takes a PromotionStrategy plus optional name/namespace and an
  optional onBack callback; the Back button only renders when onBack is given.
- HistoryPage becomes a thin dashboard wrapper that resolves the strategy from
  the store and wires router-based back navigation.
- Extension gains an Overview/History toggle and measures its wrapper offset to
  clamp the matrix to the viewport height without a stray scrollbar.
