+++ ui/components-lib/src/components/index.ts
@@ -4,6 +4,7 @@ export { default as CommitInfo } from './CommitInfo';
 export { default as HeaderBar } from './HeaderBar';
 export { default as HealthSummary } from './HealthSummary';
 export { default as HistoryNavigation } from './HistoryNavigation';
+export { default as HistoryView } from './HistoryView/HistoryView';
 export { LiveManifestView } from './LiveManifestView';
 export { default as Modal } from './Modal';
 export { StatusIcon } from './StatusIcon';
@@ -14,3 +15,4 @@ export type { StatusType } from './StatusIcon';
 export type { CardProps } from './Card';
 export type { CommitInfoProps } from './CommitInfo';
 export type { HealthSummaryProps } from './HealthSummary';
+export type { HistoryViewProps } from './HistoryView/HistoryView';
