+++ ui/extension/StrategyDropdown.scss
@@ -1,7 +1,25 @@
+.gp-controls {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 12px;
+  margin-bottom: 16px;
+}
+
+// The wrapper, not just the inner .hp, owns the viewport clamp. --hp-offset-top
+// is its measured distance from the top of the viewport (set in
+// AppViewExtension); capping the wrapper to the remaining height with
+// overflow: hidden contains the inner .hp even if it computes a sub-pixel too
+// tall, so no stray outer scrollbar appears.
+.gp-history-wrapper {
+  max-height: calc(100vh - var(--hp-offset-top, 0px));
+  overflow: hidden;
+}
+
 .strategy-dropdown-wrapper {
   display: flex;
   justify-content: center;
-  margin-bottom: 16px;
 }
 
 .strategy-dropdown__option {
@@ -89,4 +107,43 @@
 
 .strategy-dropdown__single-value {
   color: #2d3748 !important;
+}
+
+.gp-view-toggle {
+  display: flex;
+  justify-content: center;
+  gap: 4px;
+}
+
+.gp-view-toggle__btn {
+  display: flex;
+  align-items: center;
+  min-height: 38px;
+  padding: 0 16px;
+  font-size: 0.875rem;
+  font-weight: 600;
+  color: #4a5a6a;
+  background: white;
+  border: 1px solid #ccd6dd;
+  cursor: pointer;
+
+  &:first-child {
+    border-radius: 8px 0 0 8px;
+  }
+
+  &:last-child {
+    border-radius: 0 8px 8px 0;
+    border-left: none;
+  }
+
+  &:hover {
+    border-color: #0dadea;
+    color: #0dadea;
+  }
+
+  &--active {
+    background: #e6f6ff;
+    color: #0dadea;
+    border-color: #0dadea;
+  }
 }
